invoicing_django/requirements.txt
asgiref==3.7.2
Django==4.0.10
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
PyJWT==2.8.0
pymongo==3.12.2
pytz==2023.4
sqlparse==0.2.4
tzdata==2023.4
